Move dependency detection to a separate module
